revert: restore exact WSL2 working state (commit ad7ed82)
Reverted all CUDA-related files to the exact state from commit ad7ed82
which trained stably on WSL2 (41% → 55% → 58% at epochs 1-3).

Config:
- cuDNN conv2d: enabled (IMPLICIT_PRECOMP_GEMM forward, ALGO_1 backward)
- cuDNN BN: enabled
- Shadow caches: disabled (find returns nullptr)
- CUDA relu/add offload: enabled with >100K threshold
- No grad shadow chain
- No managed memory

All debugging commits since ad7ed82 reverted. Clean slate.

+++ core/serialization/onnx_export.cpp
@@ -357,7 +357,7 @@ bool convert_flatten(ONNXContext& ctx) {
     ctx.current_shape = {ctx.current_shape[0], flat_size};
     ctx.current_input = output_name;
 
-    if (ctx.verbose) printf("  Flatten -> Flatten (size=%lld)\n", flat_size);
+    if (ctx.verbose) printf("  Flatten -> Flatten (size=%ld)\n", (long)flat_size);
     return true;
 }
 
@@ -459,7 +459,7 @@ bool export_onnx(Sequential* model, const std::string& filepath, const ONNXExpor
         printf("ONNX Export: %s\n", options.model_name.c_str());
         printf("Input shape: [");
         for (size_t i = 0; i < ctx.current_shape.size(); i++) {
-            printf("%lld%s", ctx.current_shape[i], i < ctx.current_shape.size() - 1 ? ", " : "");
+            printf("%ld%s", (long)ctx.current_shape[i], i < ctx.current_shape.size() - 1 ? ", " : "");
         }
         printf("]\n");
         printf("Converting layers:\n");
@@ -533,7 +533,7 @@ bool export_onnx(Sequential* model, const std::string& filepath, const ONNXExpor
     if (options.verbose) {
         printf("Output shape: [");
         for (size_t i = 0; i < ctx.current_shape.size(); i++) {
-            printf("%lld%s", ctx.current_shape[i], i < ctx.current_shape.size() - 1 ? ", " : "");
+            printf("%ld%s", (long)ctx.current_shape[i], i < ctx.current_shape.size() - 1 ? ", " : "");
         }
         printf("]\n");
         printf("Exported %zu nodes, %zu initializers\n", ctx.nodes.size(), ctx.initializers.size());
